+++ .vs/tasks.vs.json
@@ -1,0 +1,15 @@
+{
+  "version": "0.2.1",
+  "tasks": [
+    {
+      "taskLabel": "task-README",
+      "appliesTo": "README.md",
+      "type": "launch"
+    },
+    {
+      "taskLabel": "task-scheduler",
+      "appliesTo": "scheduler.cpp",
+      "type": "launch"
+    }
+  ]
+}
