--- Capstone-Scheduler/Capstone-Scheduler.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31729.503
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Capstone-Scheduler", "Capstone-Scheduler\Capstone-Scheduler.vcxproj", "{36DB2C87-5355-4D15-9D37-505AD16C356F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{36DB2C87-5355-4D15-9D37-505AD16C356F}.Debug|x64.ActiveCfg = Debug|x64
-		{36DB2C87-5355-4D15-9D37-505AD16C356F}.Debug|x64.Build.0 = Debug|x64
-		{36DB2C87-5355-4D15-9D37-505AD16C356F}.Debug|x86.ActiveCfg = Debug|Win32
-		{36DB2C87-5355-4D15-9D37-505AD16C356F}.Debug|x86.Build.0 = Debug|Win32
-		{36DB2C87-5355-4D15-9D37-505AD16C356F}.Release|x64.ActiveCfg = Release|x64
-		{36DB2C87-5355-4D15-9D37-505AD16C356F}.Release|x64.Build.0 = Release|x64
-		{36DB2C87-5355-4D15-9D37-505AD16C356F}.Release|x86.ActiveCfg = Release|Win32
-		{36DB2C87-5355-4D15-9D37-505AD16C356F}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7B439D1C-4D8E-4085-9E40-F88D7578FFA9}
-	EndGlobalSection
-EndGlobal
