--- Capstone-Scheduler/Capstone-Scheduler/RoomInput.csv
@@ -1,2 +1,0 @@
-2,100,8,14,,
-1,50,7,14,,
