+++ RoomInput.csv
@@ -1,0 +1,2 @@
+2,100,8,14,,
+1,50,7,14,,
