+++ sess.csv
@@ -1,0 +1,5 @@
+1,120,25,RoundTable,Wifi,1
+2,30,20,Panel,Projector Speaker,2
+3,45,20,RoundTable,Wifi,2
+4,50,40,Panel,Projector Speaker,2
+5,45,40,Panel,Projector Speaker,1
